+++ Foodieland.Application/Common/Interfaces/Authentication/IJwtTokenGenerator.cs
@@ -1,6 +1,8 @@
-namespace Foodieland.Application.Common.Interfaces;
+using Foodieland.Domain.Entities;
+
+namespace Foodieland.Application.Common.Interfaces.Authentication;
 
 public interface IJwtTokenGenerator
 {
-    string GenerateToken(Guid userId, string firstName, string lastName);
+    string GenerateToken(User user);
 }
