+++ Foodieland.Application/Services/Authentication/AuthenticationResult.cs
@@ -1,8 +1,7 @@
+using Foodieland.Domain.Entities;
+
 namespace Foodieland.Application.Services.Authentication;
 
 public record AuthenticationResult(
-    Guid Id,
-    string FirstName,
-    string LastName,
-    string Email,
+    User User,
     string Token);
